Move components out of pages directory
